Evil Gluey is finally here!
I added Evil Gluey / Yeulg, to the game with a finished movement and dash script + Material.
I also have the Lazerbeem which ill use to make the second attack.
Finally if you believe the code doesn't work the best and have an idea to boost it then go for it.